init bench test
